# docs/architecture/openapi/components/openapi.yaml
openapi: 3.0.3
info:
  title: Зеленый квартал — Учёт электроэнергии API
  version: 0.8.0
  description: Канонический REST-контракт. Обновляйте синхронно с кодом WebAPI и клиента.

servers:
  - url: /api/v1
    description: Внутренний API v1

paths:
  /health:
    get:
      operationId: getHealth
      summary: Проверка готовности сервиса
      tags: [System]
      responses:
        '200':
          description: Сервис доступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /buildings:
    get:
      operationId: listBuildings
      summary: Список домов
      tags: [Directories]
      responses:
        '200':
          description: Список домов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBuildingsResponse'
    post:
      operationId: createBuilding
      summary: Создать дом
      tags: [Directories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingRequest'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBuildingResponse'
        '400':
          description: Ошибка валидации

  /buildings/{buildingId}:
    delete:
      operationId: deleteBuilding
      summary: Удалить дом
      tags: [Directories]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Удалено
        '404':
          description: Не найдено
    put:
      operationId: updateBuilding
      summary: Обновить дом
      tags: [Directories]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBuildingRequest'
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateBuildingResponse'
        '400':
          description: Ошибка валидации
        '404':
          description: Не найдено

  /buildings/{buildingId}/apartments:
    get:
      operationId: listApartmentsWithOwners
      summary: Квартиры дома с владельцами
      tags: [Directories]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список квартир
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListApartmentsWithOwnersResponse'
        '404':
          description: Дом не найден
    post:
      operationId: createApartment
      summary: Создать квартиру в доме
      tags: [Directories]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApartmentRequest'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApartmentResponse'
        '400':
          description: Ошибка валидации
        '404':
          description: Дом не найден
        '409':
          description: Дубликат номера квартиры

  /apartments/{apartmentId}:
    put:
      operationId: updateApartment
      summary: Обновить квартиру
      tags: [Directories]
      parameters:
        - name: apartmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApartmentRequest'
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateApartmentResponse'
        '400':
          description: Ошибка валидации
        '404':
          description: Квартира не найдена
        '409':
          description: Дубликат номера квартиры
    delete:
      operationId: deleteApartment
      summary: Удалить квартиру
      tags: [Directories]
      parameters:
        - name: apartmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Удалено
        '404':
          description: Не найдено

  /apartments/{apartmentId}/owner:
    put:
      operationId: upsertApartmentOwner
      summary: Назначить или обновить владельца квартиры
      tags: [Directories]
      parameters:
        - name: apartmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertApartmentOwnerRequest'
      responses:
        '200':
          description: Сохранено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpsertApartmentOwnerResponse'
        '400':
          description: Ошибка валидации
        '404':
          description: Квартира не найдена

  /apartments/{apartmentId}/meter-readings:
    get:
      operationId: listApartmentMeterReadings
      summary: История показаний по квартире
      tags: [MeterReadings]
      parameters:
        - name: apartmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: История показаний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListApartmentMeterReadingsResponse'
        '404':
          description: Квартира не найдена
    post:
      operationId: submitMeterReading
      summary: Передать или обновить показание по квартире
      tags: [MeterReadings]
      parameters:
        - name: apartmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitMeterReadingRequest'
      responses:
        '201':
          description: Показание создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitMeterReadingResponse'
        '200':
          description: Показание обновлено за тот же период
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitMeterReadingResponse'
        '400':
          description: Ошибка валидации
        '404':
          description: Квартира не найдена

  /buildings/{buildingId}/meter-readings:
    get:
      operationId: listBuildingMeterReadings
      summary: Сводка сдачи показаний по дому за период
      tags: [MeterReadings]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: periodYear
          in: query
          required: true
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
        - name: periodMonth
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: Список квартир со статусом сдачи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBuildingMeterReadingsResponse'
        '400':
          description: Ошибка валидации параметров периода
        '404':
          description: Дом не найден

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: GQ.WebApi.WebApp

    BuildingDto:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 256
        address:
          type: string
          maxLength: 512
          nullable: true

    ApartmentWithOwnerDto:
      type: object
      required: [id, buildingId, number]
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        number:
          type: string
          maxLength: 32
        floor:
          type: integer
          nullable: true
        meterVerificationDate:
          type: string
          format: date
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        ownerFullName:
          type: string
          maxLength: 256
          nullable: true
        ownerPhone:
          type: string
          maxLength: 32
          nullable: true
        lastReadingSubmittedAt:
          type: string
          format: date-time
          nullable: true
        lastReadingValue:
          type: number
          format: decimal
          nullable: true
        currentPeriodSubmitted:
          type: boolean

    OwnerDto:
      type: object
      required: [id, apartmentId, fullName]
      properties:
        id:
          type: string
          format: uuid
        apartmentId:
          type: string
          format: uuid
        fullName:
          type: string
          maxLength: 256
        phone:
          type: string
          maxLength: 32
          nullable: true

    CreateBuildingRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 256
        address:
          type: string
          maxLength: 512
          nullable: true

    CreateBuildingResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: '#/components/schemas/BuildingDto'

    UpdateBuildingRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 256
        address:
          type: string
          maxLength: 512
          nullable: true

    UpdateBuildingResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: '#/components/schemas/BuildingDto'

    ListBuildingsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BuildingDto'

    CreateApartmentRequest:
      type: object
      required: [number]
      properties:
        number:
          type: string
          maxLength: 32
        floor:
          type: integer
          nullable: true
        meterVerificationDate:
          type: string
          format: date
          nullable: true

    UpdateApartmentRequest:
      type: object
      required: [number]
      properties:
        number:
          type: string
          maxLength: 32
        floor:
          type: integer
          nullable: true
        meterVerificationDate:
          type: string
          format: date
          nullable: true

    UpdateApartmentResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: '#/components/schemas/ApartmentWithOwnerDto'

    CreateApartmentResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: '#/components/schemas/ApartmentWithOwnerDto'

    ListApartmentsWithOwnersResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApartmentWithOwnerDto'

    UpsertApartmentOwnerRequest:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
          maxLength: 256
        phone:
          type: string
          maxLength: 32
          nullable: true

    UpsertApartmentOwnerResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: '#/components/schemas/OwnerDto'

    MeterReadingDto:
      type: object
      required: [id, apartmentId, periodYear, periodMonth, value, submittedAt]
      properties:
        id:
          type: string
          format: uuid
        apartmentId:
          type: string
          format: uuid
        periodYear:
          type: integer
        periodMonth:
          type: integer
          minimum: 1
          maximum: 12
        value:
          type: number
          format: decimal
        submittedAt:
          type: string
          format: date-time

    SubmitMeterReadingRequest:
      type: object
      required: [periodYear, periodMonth, value]
      properties:
        periodYear:
          type: integer
          minimum: 2000
          maximum: 2100
        periodMonth:
          type: integer
          minimum: 1
          maximum: 12
        value:
          type: number
          format: decimal
          minimum: 0

    SubmitMeterReadingResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: '#/components/schemas/MeterReadingDto'

    BuildingMeterReadingStatusDto:
      type: object
      required: [apartmentId, apartmentNumber, submitted]
      properties:
        apartmentId:
          type: string
          format: uuid
        apartmentNumber:
          type: string
        ownerFullName:
          type: string
          maxLength: 256
          nullable: true
        readingValue:
          type: number
          format: decimal
          nullable: true
        submitted:
          type: boolean

    ListBuildingMeterReadingsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BuildingMeterReadingStatusDto'

    ListApartmentMeterReadingsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MeterReadingDto'